break up instrumenter so preview code won't be injected into the manager
