+++ code/addons/interactions/src/manager.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react';
 import type { Combo } from '@storybook/manager-api';
 import { addons, Consumer, types, useAddonState } from '@storybook/manager-api';
 import { AddonPanel, Badge, Spaced } from '@storybook/components';
-import { CallStates } from '@storybook/instrumenter';
+import { CallStates } from '@storybook/instrumenter/events';
 import { ADDON_ID, PANEL_ID } from './constants';
 import { Panel } from './Panel';
 import { TabIcon } from './components/TabStatus';
