+++ code/lib/instrumenter/src/index.ts
@@ -1,3 +1,4 @@
-export { EVENTS, instrument } from './instrumenter';
+export { instrument } from './instrumenter';
+export { EVENTS, CallStates } from './events';
 
 export * from './types';
