+++ code/lib/instrumenter/src/instrumenter.ts
@@ -12,17 +12,7 @@ import {
 import { global } from '@storybook/global';
 
 import type { Call, CallRef, ControlStates, LogItem, Options, State, SyncPayload } from './types';
-import { CallStates } from './types';
-
-export const EVENTS = {
-  CALL: 'storybook/instrumenter/call',
-  SYNC: 'storybook/instrumenter/sync',
-  START: 'storybook/instrumenter/start',
-  BACK: 'storybook/instrumenter/back',
-  GOTO: 'storybook/instrumenter/goto',
-  NEXT: 'storybook/instrumenter/next',
-  END: 'storybook/instrumenter/end',
-};
+import { CallStates, EVENTS } from './events';
 
 type PatchedObj<TObj> = {
   [Property in keyof TObj]: TObj[Property] & { __originalFn__: PatchedObj<TObj> };
